fix: update glob to version 8 and fix broken tests
In this release glob stopped accepting backward slashes on it's patterns, since `path.join` wil lreturn backward slashes on windows it they have to be replaced by a forward slash which is the posix separator.

I'm not entirely sure about all the consequences of this
change but all the tests seem to be passing and everything should work normally AFAIK since windows accept both backward and forward slashes as path separator.

Closes #281 since glob is already being updated here.